"Last Stand", "Second Wind", new "Frost Fire". Also tweaked enemy numbers

+++ header/ComponentHelper.h
@@ -72,9 +72,9 @@ public:
 
 // Defensive Stats
 	// Get max health
-	float GetMaxHealth() const;
+	int64_t GetMaxHealth() const;
 	// Get current health
-	float GetHealth() const;
+	int64_t GetHealth() const;
 	// Get a value from 0 to 1 representing the current health of the entity
 	float GetHealthFraction() const;
 	// Update the entity's bonus health
